- Aggiunta la possibilità di raggiungere un punto preciso tramite un movimento     in posizione.   - Aggiunto lo stato IN_POSIZIONE
  - Spostate le funzioni per la gestione degli eventi asincroni in
    CANOpenMasterError.h
  - Aggiunta la gestione di righe errate nei file .mot

  - La scrittura delle posizioni sulla pipe è stata spostata in un thread
    separato

  - Eliminate un po' di funzioni non più utilizzate

+++ file_parser.h
@@ -27,7 +27,7 @@ struct table_data
   int read_pointer; /**< dove sono arrivato a leggere */
   int count; /**< numero di elementi presenti nel buffer */
 
-  int cursor_position; /**< dove sono arrivato nel file delle posizioni */
+  long cursor_position; /**< dove sono arrivato nel file delle posizioni */
   int end_reached; /**< indica se non ci sono più punti da inserire nella tabella */
 
   pthread_mutex_t table_mutex; /**< sincro tra diversi thread */
